moved the aerodynamic length scale
